Added a ton of stuff. Not sure why Mars is not showing up. Trying to fix it.
