+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Scene from './components/Background/Scene';
+import Scene from './components/Scene';
 
 function App() {
     return (
